(#10) Introduce CPlatform abstract class
